requirements.txt
django = "~=3.1.0"
djangorestframework = "==3.12.0"
pillow = "*"
djoser = "*"